requirements.txt
langchain>=0.1.0
langchain-core>=0.1.0
langchain-community>=0.1.0
PyPDF2>=3.0.0
sentence-transformers>=2.2.2
faiss-cpu>=1.7.4
gpt4all>=2.0.0
llama-cpp-python>=0.2.0
unstructured>=0.10.0
pymupdf>=1.23.0
streamlit>=1.31.0
huggingface-hub>=0.19.0
python-dotenv>=0.19.0
